Switch to lower-level Slack API
The hanu-based chatbot would lose its connection after less than a day
of time. Rather than try to get to the very bottom of why that is, I'm
switching this to a more full-featured but lower-level Slack package. I
wanted a little more access to the full Slack API anyway.